move TailTasks into `task_executor` to avoid crate cycle
[ci skip-build-wheels]
