+++ src/rust/engine/process_execution/src/lib.rs
@@ -37,15 +37,15 @@ use async_trait::async_trait;
 use concrete_time::{Duration, TimeSpan};
 use deepsize::DeepSizeOf;
 use fs::{DirectoryDigest, RelativePath, EMPTY_DIRECTORY_DIGEST};
-use futures::future::{try_join_all, BoxFuture};
+use futures::future::try_join_all;
 use futures::try_join;
 use hashing::Digest;
 use itertools::Itertools;
-use parking_lot::Mutex;
 use protos::gen::build::bazel::remote::execution::v2 as remexec;
 use remexec::ExecutedActionMetadata;
 use serde::{Deserialize, Serialize};
 use store::{SnapshotOps, Store, StoreError};
+use task_executor::TailTasks;
 use workunit_store::{in_workunit, Level, RunId, RunningWorkunit, WorkunitStore};
 
 pub mod bounded;
