+++ src/rust/engine/task_executor/Cargo.toml
@@ -9,6 +9,8 @@ publish = false
 arc-swap = "1.5"
 futures = "0.3"
 lazy_static = "1"
+log = "0.4"
+parking_lot = "0.12"
 stdio = { path = "../stdio" }
 tokio = { version = "1.21", features = ["rt-multi-thread"] }
 workunit_store = { path = "../workunit_store" }
